[FLINK-23284] Reset the TARGET_COUNT_REACHED_LATCH before each execution of the job
The Level.Iteration is not enough to clear up the LATCH before every run of the readFileSplit method. There might be multiple calls to the benchmarked method within a single iteration. If it is the case, the second invocation did not have the latch properly reset, which resulted in all records being emitted in the finishing stage of the ContinuousFileReadingOperator. This led to testing a different behaviour than the envisioned.